+++ src/main/java/com/telepathicgrunt/the_bumblezone/items/EmptyBucketDispenseBehavior.java
@@ -1,6 +1,7 @@
 package com.telepathicgrunt.the_bumblezone.items;
 
 import com.telepathicgrunt.the_bumblezone.blocks.BzBlocks;
+import com.telepathicgrunt.the_bumblezone.mixin.DefaultDispenseItemBehaviorInvoker;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.DispenserBlock;
 import net.minecraft.dispenser.DefaultDispenseItemBehavior;
@@ -43,7 +44,15 @@ public class EmptyBucketDispenseBehavior extends DefaultDispenseItemBehavior {
             else
                 stack = new ItemStack(BzItems.SUGAR_WATER_BUCKET.get());
         } else {
-            return DEFAULT_EMPTY_BUCKET_DISPENSE_BEHAVIOR.dispense(source, stack);
+            // If it instanceof DefaultDispenseItemBehavior, call dispenseStack directly to avoid
+            // playing particles and sound twice due to dispense method having that by default.
+            if(DEFAULT_EMPTY_BUCKET_DISPENSE_BEHAVIOR instanceof DefaultDispenseItemBehavior) {
+                return ((DefaultDispenseItemBehaviorInvoker)DEFAULT_EMPTY_BUCKET_DISPENSE_BEHAVIOR).invokeDispenseStack(source, stack);
+            }
+            else {
+                // Fallback to dispense as someone chose to make a custom class without dispenseStack.
+                return DEFAULT_EMPTY_BUCKET_DISPENSE_BEHAVIOR.dispense(source, stack);
+            }
         }
 
         return stack;